Create Max Sum SubSequence in an Array
Given an array A of N positive integers. Find the sum of maximum sum ""non-decreasing"" sub sequence of the given array.
Example : Given array 
1 101 2 3 100 4 5
The Maximum Sum non decreasing Sub Sequence will be
106 : 1+2+3+100

Approach : Dynamic Programming 
Maintain an array which contains the maximum sum of the non-decreasing sub sequence at each element (i.e from index 0 to index of the element) 
For the array 1 101 2 3 100 4 5 the array of max sum sub sequence would be 
1st iteration : 1 0 0 0 0 0 0 0
2nd iteration: 1 102 0 0 0 0 0 0 // Since (current elem)101>1 so max sum at 101 will be 101 + 1
3rd iteration : 1 102 3 0 0 0 0 0 // Since (current elem )2<101 so max sum cant be 102 +  2 . Since 2 > 1 max sum can be 2+1 
and so on .
The maximum value in the max sum sub sequence array will be the answer.